Minor changes to Q32 sqrt and var functions -- var overflows for some cases

+++ src/StatisticsFunctions/kernels/plp_var_q16s_xpulpv2.c
@@ -81,10 +81,10 @@ void plp_var_q16s_xpulpv2(
   int16_t mean;
 
   plp_mean_i16(pSrc, blockSize, &mean);
-  square_of_mean = ((int32_t)mean * (int32_t)mean) >> deciPoint;
+  square_of_mean = (mean * mean) >> deciPoint;
 
   plp_power_q16(pSrc, blockSize, deciPoint, &square_of_values);
   
-  *pRes = (int16_t)(square_of_values/blockSize - square_of_mean);
+  *pRes = (square_of_values/blockSize - square_of_mean);
 
 }
